Drop GHC installation via apt
The apt repo hosting these packages has been unmaintained for quite a
while, and we've excluded ~/.stack/programs from the cache long time
ago, so we can just let stack handle the installation. This step has
been failing for a while anyway, as there's no ghc-9.0.2 in the apt
repo.